Remove example data files and rename list component to statistics
