+++ ui/reval-web/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { CommonModule } from '@angular/common';
-import { ListComponent } from './list/list.component';
+import { ListComponent } from './statistics/statistics.component';
 import { EvaluationComponent } from './evaluation/evaluation.component';
 
 const routes: Routes = [
