+++ ui/reval-web/src/app/app.module.ts
@@ -15,7 +15,7 @@ import { MatBadgeModule } from '@angular/material/badge';
 import { EvaluationComponent } from './evaluation/evaluation.component';
 import { AuthConfigModule } from './auth/auth-config.module';
 import { UserprofileComponent } from './userprofile/userprofile.component';
-import { ListComponent } from './list/list.component';
+import { ListComponent } from './statistics/statistics.component';
 import { MatListModule } from '@angular/material/list';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
 import { Configuration } from 'src/openapi-client/evaluationapi';
