--- ui/reval-web/src/app/list/data.ts
@@ -1,14 +1,0 @@
-export var evaluations = [
-    {
-        "name": "negative",
-        "value": 38
-    },
-    {
-        "name": "unanswered",
-        "value": 20
-    },
-    {
-        "name": "positive",
-        "value": 360
-    }
-];
