--- ui/reval-web/src/app/list/exampleRating.json
@@ -1,39 +1,0 @@
-{
-    "scenarios": [
-        {
-            "id": "7d1e8d9a-f0c6-4f12-86d4-6956e1b47e01",
-            "name": "Szenario Vicuna",
-            "description": "Das ist das erste tolle Szenario mit den Antworten aus Llama-2-13b-chat-hf_responses_rated_by_gpt-3.5-turbo.json",
-            "totalResponseCount": 400,
-            "progressStatistics": [
-                {
-                    "name": "rated",
-                    "value": 70
-                },
-                {
-                    "name": "unrated",
-                    "value": 330
-                }
-            ],
-            "resultStatistic": [
-                {
-                    "name": "positive",
-                    "value": 12
-                },
-                {
-                    "name": "negative",
-                    "value": 46
-                },
-                {
-                    "name": "neutral",
-                    "value": 12
-                }
-            ],
-            "ratingScore": {
-                "min": -1,
-                "value": 0.8,
-                "max": 1
-            }
-        }
-    ]
-}
