+++ ui/reval-web/src/app/statistics/statistics.component.html
