+++ ui/reval-web/src/app/statistics/statistics.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { ListComponent } from './list.component';
+import { ListComponent } from './statistics.component';
 
 describe('ListComponent', () => {
   let component: ListComponent;
